mingw64 build cleanup
Fixes to allow build to run to completion in the Fedora x86_64 mingw64
cross-compilation environment.  Mostly (but not exclusively) in
exec.c for libpcp.

Also fixes to address many of the compilation warnings ... we're down
to less than 50 now!

Lots of these are because the mingw64 headers define getpid() to
return an int, not a pid_t ... leading to (pid_t)getpid() casts
that are no-ops for everyone else.

+++ src/pmdas/windows/open.c
@@ -673,7 +673,7 @@ windows_visit_metric(pdh_metric_t *pmp, pdh_metric_visitor_t visitor)
 	    if ((pmp->vals = (pdh_value_t *)realloc(pmp->vals, size)) == NULL) {
 		__pmNotifyErr(LOG_ERR, "windows_open: Error: values realloc "
 				   "(%d x %d) failed @ metric %s [%s]: ",
-				pmp->num_vals, sizeof(pdh_value_t),
+				pmp->num_vals, (int)sizeof(pdh_value_t),
 				pmIDStr(pmp->desc.pmid), p);
 		pmp->num_alloc = 0;
 		return -1;
